+++ sentry-sidekiq/example/error_worker.rb
@@ -10,6 +10,8 @@ end
 class ErrorWorker
   include Sidekiq::Worker
 
+  sidekiq_options retry: 0
+
   def perform
     1 / 0
   end